Extract MapId/TOWN_LABELS into a dependency-free module
loader.ts needs MapId for NpcScheduleEntry's upcoming 'travel' location
type, but hubWorldFactory.ts (which previously declared MapId inline)
already imports from loader.ts — importing back would cycle. Moves
MapId and the map editor's TOWN_LABELS/townLabel (previously duplicated
in entityRefs.ts) into a new zero-dependency mapIds.ts that both can
import from. Pure refactor: both files re-export from the new module,
so all 13 existing importers are unaffected.
